Detect the path to ltmain.sh at build time and set conflicts appropriately
NOTE: this is rebased to include the later changes in debian-sid that
improved the detection and aesthetics.

Since version 2.2 libtool ships ltmain.sh under /usr/share/libtool/config,
which lead to broken symlinks. debian/rules will now try to find the right
path at build time and add a conflict on pre 2.2 versions of libtool if
the path is usr/share/libtool/config, or a conflict on version 2.2 or higher
if the path is usr/share/libtool, to avoid issues when backporting.

note, this also contains changes to libtool related stuff in the patch
033-we_WANT_libtool.patch which was in seperate commits in unstable:

upstream ships an out of date version of libtool.  this ensures that
we build against an up-to-date version of libtool by running libtoolize
as part of our build process (this is called indirectly via ./buildconf.sh
from debian/rules)

note that we don't touch the libtool.m4 that they ship here, and this file
gets included in the build process as part of the phpize stuff.  however,
this is solved in ./debian/rules where it's overwritten with a symlink.