Detect the path to ltmain.sh at build time and set conflicts appropriately
Since version 2.2 libtool ships ltmain.sh under /usr/share/libtool/config,
which lead to broken symlinks. debian/rules will now try to find the right
path at build time and add a conflict on pre 2.2 versions of libtool if
the path is usr/share/libtool/config, or a conflict on version 2.2 or higher
if the path is usr/share/libtool, to avoid issues when backporting.

Thanks to Martin Meredith for the report, closes: #524163