update our libtool patching to be a little cleaner and smarter
upstream ships an out of date version of libtool.  this ensures that
we build against an up-to-date version of libtool by running libtoolize
as part of our build process (this is called indirectly via ./buildconf.sh
from debian/rules)

note that we no longer touch the libtool.m4 that they ship here, and this file
gets included in the build process as part of the phpize stuff.  however,
this is solved in ./debian/rules where it's overwritten with a symlink.