remove binary package for php5-mhash which is now built-in
the mhash extension has been removed in php5.3, replaced by built in
hash extension, which provides a BC layer for the mhash functions.

the --with-mhash configure flag is also moved from the apache2 configure
options to the common configure options, so that it will be built-in to
all php sapi's.