Switch to automake (>= 1.11) | automake1.11, depend on autoconf >= 2.63
From http://php.net/svn.php:

autoconf, automake and libtool information

There seem to be problems with libtool 1.4.2. It is suggested that you use libtool 1.4, along with autoconf 2.13 and automake 1.4. You should also ensure that autoconf, automake and libtool are installed in the same directory. libtool 1.5 will not work.

The following combinations are known to work:

autoconf 2.13, automake 1.4 and libtool 1.4.3
autoconf 2.13, automake 1.5 and libtool 1.4.3
autoconf 2.63, automake 1.11 and libtool 2.2.6 (with many warnings)

Closes: #549148