Dynamically determine maxlifetime if possible.
Incorporate suggestions from #504053 to use the cmdline php if available
to determine the maxlifetime value, but retain the previous logic as well
in case php5-cli is not installed.  Also, add a Recommends on php5-cli to
libapache2-mod-php5 to make it more likely that it is installed.

Closes: #504053
Thanks: Chris Butler <chrisb@debian.org>