Put all licences in debian/copyright.
It is a policy requirement that we have a complete debian/copyright in the
source package (not only in the binary packages). The simplest way to
assure this is to just put everything together statically. It's not like
that the PHP and Zend licence change very often, and other licences
are already hardcoded in copyright.header.